# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AI Social Media Management Platform API
  version: 0.1.0
  description: Initial REST contract for the enterprise social media management SaaS.
servers:
  - url: http://localhost:4000/api
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /safety/policies:
    get:
      tags: [safety]
      summary: List AI safety policies
      parameters:
        - in: query
          name: workspaceId
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Safety policies
    post:
      tags: [safety]
      summary: Create an AI safety policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSafetyPolicy"
      responses:
        "201":
          description: Created safety policy
  /safety/evaluate:
    post:
      tags: [safety]
      summary: Evaluate content against workspace AI safety policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluateContentSafety"
      responses:
        "201":
          description: Safety check with optional moderation queue item
  /safety/checks:
    get:
      tags: [safety]
      summary: List content safety checks
      parameters:
        - in: query
          name: workspaceId
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Content safety checks
  /safety/moderation-queue:
    get:
      tags: [safety]
      summary: List AI moderation queue items
      parameters:
        - in: query
          name: workspaceId
          schema:
            type: string
            format: uuid
        - in: query
          name: status
          schema:
            type: string
            enum: [open, approved, rejected, resolved]
      responses:
        "200":
          description: Moderation queue items
  /safety/moderation-queue/{id}/resolve:
    post:
      tags: [safety]
      summary: Resolve an AI moderation queue item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveModerationItem"
      responses:
        "201":
          description: Resolved moderation queue item
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
...
    CreateSafetyPolicy:
      type: object
      required: [workspaceId, name]
      properties:
        workspaceId:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 180
        rules:
          type: object
          properties:
            blockedTerms:
              type: array
              items:
                type: string
            requiredDisclosures:
              type: array
              items:
                type: string
            industry:
              type: string
            maxRiskScore:
              type: number
              minimum: 0
              maximum: 1
    EvaluateContentSafety:
      type: object
      required: [workspaceId, text]
      properties:
        workspaceId:
          type: string
          format: uuid
        text:
          type: string
          minLength: 1
          maxLength: 10000
        source:
          type: string
          enum: [ai_generation, manual, post_review]
        sourceEntityId:
          type: string
          format: uuid
    ResolveModerationItem:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected, resolved]
        resolutionNote:
          type: string
          maxLength: 2000
...